Update 22/11/2024 num1
Coding of the TIF model. Started off with finishing the dataset class that works with the combined TIF image format. This class will return an image with the combination of bands of the choice between

RGB ( Band 3,4,5 )

RGB-NIR ( Band 3,4,5,8 )

RGB-SWIR ( Band 3,4,5,11,12 )

RGB-NIR-SWIR ( Band 3,4,5,8,11,12 )

All bands ( Band 1-12 )

For now no transformations were set in place, due to the reason that normal transformations are designed for PIL images and it won’t work with multispectral imagery.

The DataLoader was developed together with the Model class. The model class received significant changes as opposed to the RGB version including the change of the first layer to support multiple bands. ResNet18 is still used as the primary model for testing purposes.

Finally the training phase was developed, a number of experiments were done to check how the models are performing with the 0.5% subset of BigEarthNet