Update 30/11/2024 - Replication and testing update
Massive changes to the codebase. Added a main.py where the user is able to select what model he wants, what band combination he wants and also if the model is pre-trained or not. Made changes to a large number of files to accommodate these changes.

Started my shift to the testing stage. Verified the testing function and also started creating functions to create a confusion matrix and also to predict based on a given image.