Bug Issue and Testing update
Discovery of an issue within the data loaders. The images are being converted to a tensor and transformed accordingly however the labels seem to be loaded incorrectly.

Analysis was performed and found the issue to be in the get_label() method within the dataset class. This was fixed accordingly.

Also continued on the test functions and ensured that they work accordingly. This included updating the testing 1 image function, adding a new function for the uac curve and also updating the heatmap