Experiment Details Update
Users are able to see all the experiments and press on any one to view detailed metrics of the experiments.

Further enhancements to the UI and how everything is placed and looks so that the user is able to accuaretly and efficiently view the metrics and visualisations effectively.

Also added another custom callback that adds logs based on training and testing in real time.