Map update - API calls from Sentinel Hub
The users are interact with a map provided from ESRI world imagery. Using Sentinel Hub's API, when the user clicks on a particular location on the map a 120x120 tile is taken as a TIFF file at 12 spectral bands - the same as BigEarthNet, which then is saved within the static folder of the web folder.